requirements.txt
# Core Dependencies
langchain==0.1.20
langchain-community==0.0.38
langgraph==0.0.55
langchain-openai==0.1.6
langchain-anthropic==0.1.11

# LLM Support
openai>=1.14.0
anthropic>=0.21.0
ollama>=0.1.7

# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.0
streamlit==1.32.0
gradio==4.19.0

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Task Queue
celery[redis]==5.3.6

# Literature Collection
arxiv==2.1.0
scholarly==1.7.11
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Data Processing & Analysis
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
scipy==1.12.0
scikit-learn==1.4.1.post1
statsmodels==0.14.1
plotly==5.19.0

# Advanced Statistical Analysis
pingouin==0.5.4
lifelines==0.28.0

# Document Processing
pypdf==4.0.2
PyPDF2==3.0.1
pdfplumber==0.10.4
python-docx==1.1.0
python-pptx==0.6.23
pdfminer.six==20231228
tabula-py==2.9.0

# LaTeX Support
pylatexenc==2.10
latex2mathml==3.76.0

# NLP & Text Processing
nltk==3.8.1
spacy==3.7.4
transformers==4.38.2
sentence-transformers==2.5.1
textstat==0.7.3
language-tool-python==2.7.1

# Citation Management
bibtexparser==1.4.1
habanero==1.2.6
crossref-commons==0.0.7

# Image Processing & Computer Vision
Pillow==10.2.0
opencv-python==4.9.0.80
pytesseract==0.3.10
pdf2image==1.17.0

# Code Analysis
pygments==2.17.2
radon==6.0.1
ast-grep==0.5.1

# Utilities
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-multipart==0.0.9
aiohttp==3.9.3
httpx==0.27.0
tenacity==8.2.3
tqdm==4.66.2
rich==13.7.0

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.98.9
pytest-xdist==3.5.0

# Code Quality
black==24.2.0
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit==3.6.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.11

# Proxy Support (for Google Scholar)
requests[socks]==2.31.0
PySocks==1.7.1

# Additional Tools
jinja2==3.1.3
pyyaml==6.0.1
jsonschema==4.21.1
